Took some look at "code analysis" tool output: - Removed conversion from const char* as returned by c_str() to std::string (creates an unnecessary string copy) Solved by directly passing the string. - Parameters 'p' & 'n' were passed by value. Passed as a const reference which is usually faster and recommended in C++. - Variable 'zscale' was assigned a value that is never used. - Initialized variable: Func. - Removed (some) unused variables. - Removed unused private member cloudtexidx. - Initialized members: texname, RenderParam. - Initialized members of class D3D9Client. - Optimized Record/Playback label code in D3D9Client::clbkRenderScene(). - Fixed possible access violation when using 'hNew' at D3D9Pad::TexChange().
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@999 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/D3D9Client.h
@@ -1270,7 +1270,7 @@ private:
 	Sketchpad *pItemsSkp;
 
 	DWORD loadd_x, loadd_y, loadd_w, loadd_h;
-	WORD LabelPos;
+	int LabelPos;
 
 }; // class D3D9Client
